+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Support PodDisruptionBudgets ([#625]).
 - Support new versions 2.8.2, 3.4.1, 3.5.1 ([#627]).
 - Document internal clusterId check ([#631]).
+- Support graceful shutdown ([#635]).
 
 ### Changed
 
@@ -32,6 +33,7 @@ All notable changes to this project will be documented in this file.
 [#625]: https://github.com/stackabletech/kafka-operator/pull/625
 [#627]: https://github.com/stackabletech/kafka-operator/pull/627
 [#631]: https://github.com/stackabletech/kafka-operator/pull/631
+[#635]: https://github.com/stackabletech/kafka-operator/pull/635
 
 ## [23.7.0] - 2023-07-14